renamed AsyncContainer -> AsyncPlaceholder
