+++ client/src/components/CodeViewer.tsx
@@ -4,7 +4,7 @@ import { basename, extname } from 'path'
 import React from 'react'
 import { FileType, useGetAnswerFileQuery } from '../generated/graphql'
 import { isBinaryContent, numberOfLines, sliceLines } from '../services/file'
-import AsyncPlaceholder from './AsyncContainer'
+import AsyncPlaceholder from './AsyncPlaceholder'
 import Centered from './Centered'
 import ReviewEditor from './code/ReviewEditor'
 import SyntaxHighlighter from './code/SyntaxHighlighter'
