+++ client/src/pages/answer/AnswerPage.tsx
@@ -9,7 +9,7 @@ import { useCallback, useContext, useEffect, useState } from 'react'
 import AnswerBlocker from '../../components/AnswerBlocker'
 import AnswerFileBrowser from '../../components/AnswerFileBrowser'
 import ArchiveDownload from '../../components/ArchiveDownload'
-import AsyncPlaceholder from '../../components/AsyncContainer'
+import AsyncPlaceholder from '../../components/AsyncPlaceholder'
 import FileImport from '../../components/FileImport'
 import useMomentReached from '../../hooks/useMomentReached'
 import { useServerNow } from '../../hooks/useServerTimeOffset'
