+++ client/src/pages/assignment/AssignmentPage.tsx
@@ -22,7 +22,7 @@ import { Route, Switch, useHistory, useRouteMatch } from 'react-router-dom'
 import { debounce } from 'ts-debounce'
 import ArchiveDownload from '../../components/ArchiveDownload'
 import AssignmentStatusTag from '../../components/AssignmentStatusTag'
-import AsyncPlaceholder from '../../components/AsyncContainer'
+import AsyncPlaceholder from '../../components/AsyncPlaceholder'
 import Authorized from '../../components/Authorized'
 import { createBreadcrumb } from '../../components/DefaultLayout'
 import EditableTitle from '../../components/EditableTitle'
