+++ client/src/pages/task/TaskConfigurationPage.tsx
@@ -21,7 +21,7 @@ import { JSONSchema6 } from 'json-schema'
 import React, { useState } from 'react'
 import ReactMarkdown from 'react-markdown'
 import { Link } from 'react-router-dom'
-import AsyncPlaceholder from '../../components/AsyncContainer'
+import AsyncPlaceholder from '../../components/AsyncPlaceholder'
 import EditableMarkdown from '../../components/EditableMarkdown'
 import CodefreakDocsLink from '../../components/CodefreakDocsLink'
 import JsonSchemaEditButton from '../../components/JsonSchemaEditButton'
